pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "growth-tools"
version = "0.1.0"
description = "Lead capture from Reddit, Discord and GitHub with hybrid LLM scoring and outreach drafts."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["growth", "leads", "reddit", "discord", "scoring", "outreach", "saas"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "praw>=7.7.0",
    "openai>=1.52.0",
    "supabase>=2.0.0",
    "discord.py>=2.3.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
growth-reddit  = "growth_tools.run_reddit:main"
growth-api     = "growth_tools.run_api:main"

[tool.hatch.build.targets.wheel.sources]
"src" = "growth_tools"

[project.urls]
Repository = "https://github.com/YOUR_ORG/growth-tools"
